requirements.txt
Django==1.7.1
GnuPGInterface==0.3.2
GraphLab-Create==1.2
amqp==1.4.6
anyjson==0.3.3
apt-xapian-index==0.44
argparse==1.2.1
awscli==1.6.2
backports.ssl-match-hostname==3.4.0.2
bcdoc==0.12.2
billiard==3.3.0.19
boto==2.33.0
botocore==0.73.0
celery==3.1.17
chardet==2.0.1
colorama==0.2.5
command-not-found==0.2.44
decorator==3.4.0
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
django-celery==3.1.16
docutils==0.12
jmespath==0.5.0
kombu==3.0.24
language-selector==0.1
librato-metrics==0.4.9
mixpanel-py==3.1.1
prettytable==0.7.2
pyasn1==0.1.7
python-apt==0.8.3ubuntu7
python-dateutil==2.3
python-debian==0.1.21ubuntu1
pytz==2014.10
redis==2.10.3
requests==2.3.0
rsa==3.1.2
six==1.8.0
tornado==3.2.1
ufw==0.31.1-1
wsgiref==0.1.2
